Adding support for Private DNS. Ensuring we only use official Amazon tools.
1) This commit removes quite a lot of code and cleans up a lot. Now we only use
official Amazon toolsets, and the code is much cleaner and easier to read.

2) This commit also adds support for private hosted zones - For use in a VPC.
When a private DNS is specified, we will create an A record with the instances
internal IPV4 address. If it is not specified, we will create a CNAME record
with the instances public hostname as before.

Squashed commit of the following:

commit 805691ec4794bde4fc9524732ca52af50474589c
Author: TJ Biddle <tjbiddle@inflection.com>
Date:   Tue Dec 2 13:13:35 2014 -0500

    Adding README.

commit a6fe8f7a2f4ce33572ecbf232b606e22c5547cca
Author: TJ Biddle <tjbiddle@inflection.com>
Date:   Tue Dec 2 12:55:29 2014 -0500

    Adding support for private DNS

commit 4d78898a6ff1f9cfd67e72c0977314462208dc17
Author: TJ Biddle <tjbiddle@inflection.com>
Date:   Tue Dec 2 11:22:10 2014 -0500

    Work in progress.